line_webhook_demo/requirements.txt
line-bot-sdk==3.14.2
functions-framework==3.8.2
google-cloud-discoveryengine==0.12.3
google-cloud-storage==2.18.2
google-cloud-datastore==2.20.2
vertexai==1.71.1
google-cloud-dialogflow-cx==1.36.0
google-cloud-speech
aiohttp>=3.10.9
yarl>=1.12.0
frozenlist==1.3.1
pandas